Convert all the tests in `Pest` (#3)
* refactor: move cariche famiglie aziende test into pest

* test: move to pest syntax

* test: move tests into UNit into the specific folder

* test: passing all the pest tests
