+++ phpunit.xml
@@ -6,12 +6,6 @@
     </include>
   </coverage>
   <testsuites>
-    <testsuite name="Http">
-      <directory suffix="Test.php">./tests/Http</directory>
-    </testsuite>
-    <testsuite name="Unit">
-      <directory suffix="Test.php">./tests/Unit</directory>
-    </testsuite>
     <testsuite name="Biblioteca">
       <directory suffix="Test.php">./tests/Biblioteca</directory>
     </testsuite>
@@ -20,6 +14,12 @@
     </testsuite>
     <testsuite name="Scuola">
       <directory suffix="Test.php">./tests/Scuola</directory>
+    </testsuite>
+    <testsuite name="Officina">
+      <directory suffix="Test.php">./tests/Officina</directory>
+    </testsuite>
+    <testsuite name="AdminSys">
+      <directory suffix="Test.php">./tests/AdminSys</directory>
     </testsuite>
   </testsuites>
   <php>
