+++ tests/Pest.php
@@ -1,13 +1,16 @@
 <?php
 
 use App\Admin\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Artisan;
 
+use Spatie\Permission\Exceptions\UnauthorizedException;
 use Tests\TestCasePest;
 use function Pest\Laravel\actingAs;
 
 
-uses(TestCasePest::class)->in('Biblioteca', 'Scuola', 'Popolazione');
+uses(TestCasePest::class)->in('Biblioteca', 'Scuola', 'Popolazione', 'Officina', 'AdminSys');
 
 
 function login(User $user = null): User
@@ -19,4 +22,18 @@ function login(User $user = null): User
     actingAs($user);
 
     return $user;
+}
+
+function runMiddleware($middleware, $permission) {
+    try {
+        return $middleware->handle(
+            new Request(),
+            function () {
+                return (new Response())->setContent('<html></html>');
+            },
+            $permission
+        )->status();
+    } catch (UnauthorizedException $e) {
+        return $e->getStatusCode();
+    }
 }
