+++ tests/Popolazione/Unit/AziendaTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Tests\Unit;
+
+use Domain\Nomadelfia\Azienda\Models\Azienda;
+use Domain\Nomadelfia\Persona\Models\Persona;
+use Tests\CreatesApplication;
+use Tests\MigrateFreshDB;
+use Tests\TestCase;
+use Carbon\Carbon;
+
+
+it("assign a worker to a company", function () {
+    $azienda = Azienda::factory()->create();
+    $persona = Persona::factory()->maggiorenne()->maschio()->create();
+
+    expect($azienda->lavoratori()->count())->toBe(0);
+    expect($azienda->lavoratoriAttuali()->count())->toBe(0);
+    expect($azienda->lavoratoriStorici()->count())->toBe(0);
+    expect($persona->aziendeAttuali()->count())->toBe(0);
+    expect($persona->aziendeStorico()->count())->toBe(0);
+
+
+    $data_inizio = Carbon::now()->addYears(5)->toDatestring();
+    $persona->assegnaLavoratoreAzienda($azienda, $data_inizio);
+    expect($azienda->lavoratoriAttuali()->count())->toBe(1);
+
+    $resp = Persona::factory()->maggiorenne()->maschio()->create();
+    $resp->assegnaResponsabileAzienda($azienda, $data_inizio);
+    expect($azienda->lavoratoriAttuali()->count())->toBe(2);
+
+    $data_uscita = Carbon::now()->addYears(5)->toDatestring();
+    $persona->uscita($data_uscita);
+
+    expect($azienda->lavoratoriAttuali()->count())->toBe(1);
+});
