+++ tests/TestCasePest.php
@@ -2,14 +2,13 @@
 
 namespace Tests;
 
-use App\Admin\Models\User;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 use Symfony\Component\HttpFoundation\Request as SymfonyRequest;
 use Illuminate\Http\Request;
 
 abstract class TestCasePest extends BaseTestCase
 {
-    use CreatesApplication;
+    use CreatesApplication, MigrateFreshDB;
 
     public function createRequest($method, $uri): Request
     {
